Collections can now be heterogeneous.
That is, they can include objects of more than one type. Pass in an
array of name, constructor pairs. The restish endpoint for the
collection will be expected to return a hash with a 'type' element for
each collection item, specifying which of the types is it. The data is
passed to reset after construction as usual.