Got the redis adapter working & generalized persistence considerably.
Methodically got all the unit tests for the redis adapter working by
methodically implementing all the redis functions.
Some couch-ness in the persistence layer remains.
Moved hiredis to the optional deps because it appears not to build for
node 0.9.
Broke the couch-specific unit tests out.
Fiddled with optional properties a bit in the base model.